refactoring cards; unifying styles for landing cards
