+++ src/app/map/[monitor_id]/layout.tsx
@@ -2,7 +2,7 @@ import type { FC, ReactNode } from 'react';
 
 import dynamic from 'next/dynamic';
 
-import MonitorCard from '@/components/monitors/card';
+import MonitorHeader from '@/components/monitors/header';
 import TabsNav from '@/components/tabs-nav';
 import { ScrollArea } from '@/components/ui/scroll-area';
 
@@ -14,7 +14,7 @@ const MonitorLayout: FC<{ children: ReactNode }> = ({ children }) => {
       <section className="md:[30vw] absolute bottom-3 left-3 top-3 z-40 w-[526px] overflow-hidden bg-brand-500">
         <ScrollArea className="h-full w-full p-7.5" type="auto">
           <div className="space-y-6">
-            <MonitorCard />
+            <MonitorHeader />
             <TabsNav />
           </div>
           {children}
