+++ src/app/page.tsx
@@ -3,7 +3,7 @@ import type { NextPage, Metadata } from 'next';
 import LandingDatasets from '@/components/datasets-grid';
 import Footer from '@/components/footer';
 import Hero from '@/components/hero';
-import Prefooter from '@/components/prefooter';
+import PreFooter from '@/components/pre-footer';
 
 export const metadata: Metadata = {
   title: 'Hub - Open Earth Monitor Cyberinfrastructure',
@@ -14,7 +14,7 @@ const Hub: NextPage = () => (
   <>
     <Hero />
     <LandingDatasets />
-    <Prefooter />
+    <PreFooter />
     <Footer />
   </>
 );
