+++ src/types/monitors.d.ts
@@ -9,6 +9,11 @@ export type Monitor = {
   id: string;
   title: string;
   entity_type: 'monitor';
+  ready: boolean;
+  metadata_url: string;
+  notebooks_url: string;
+  publications: { title: string; url: string }[];
+  use_case_link: { title: string; url: string }[];
 };
 
 export type MonitorParsed = Monitor & { color: string; colorOpacity: string };
